Improved Webpack-style tilde-import support for Sass
The current solution for dealing with sass/scss imports which are
prefixed with a tilde (~) is to strip the symbol from the import line.
This breaks down when dart-sass itself starts pulling in subsequent
imports that are also tilde-prefixed.

This change introduces a simple algorithm to climb the filesystem tree
looking for the imported module/file in a node_modules directory.

This is enabled by default since it replaces behavior which is already
forced on.

This change is covered under the existing "with a Bootstrap import"
test.